Replace spatie/macroable with own-baked macro()
