+++ tests/PhlnTest.php
@@ -13,17 +13,6 @@ class PhlnTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals('foo', P::foo());
     }
 
-    public function test_it_allows_to_define_curried_macro()
-    {
-        P::curried('foo', 2, function ($a, $b) {
-            return $a + $b;
-        });
-
-        $f = P::foo(1);
-
-        $this->assertEquals(3, $f(2));
-    }
-
     public function test_adding_macro_aliases()
     {
         P::macro('hello', function (string $name) {
@@ -38,5 +27,47 @@ class PhlnTest extends \PHPUnit\Framework\TestCase
     public function test_it_returns_ref_to_macro()
     {
         $this->assertEquals(P::class.'::tap', P::ref('tap'));
+    }
+
+    public function test_it_returns_arity_of_function()
+    {
+        $this->assertEquals(1, P::arity(function ($a) {}));
+        $this->assertEquals(2, P::arity(function ($a, $b) {}));
+        $this->assertEquals(1, P::arity(function (...$a) {}));
+        $this->assertEquals(2, P::arity(function ($a, ...$b) {}));
+    }
+
+    public function test_curry_with_known_arity()
+    {
+        $sum = function ($a, $b, $c) {
+            return $a + $b + $c;
+        };
+
+        $expected = $sum(1, 2, 3);
+        $g = P::curryN(3, $sum);
+
+        $this->assertEquals($expected, $g(1, 2, 3));
+        $this->assertEquals($expected, $g(1)(2, 3));
+        $this->assertEquals($expected, $g(1, 2)(3));
+        $this->assertEquals($expected, $g(1)(2)(3));
+
+        $this->assertEquals($expected, P::curryN(3, $sum, [1, 2, 3]));
+    }
+
+    public function test_curry()
+    {
+        $sum = function ($a, $b, $c) {
+            return $a + $b + $c;
+        };
+
+        $expected = $sum(1, 2, 3);
+        $g = P::curry($sum);
+
+        $this->assertEquals($expected, $g(1, 2, 3));
+        $this->assertEquals($expected, $g(1)(2, 3));
+        $this->assertEquals($expected, $g(1, 2)(3));
+        $this->assertEquals($expected, $g(1)(2)(3));
+
+        $this->assertEquals($expected, P::curry($sum, [1, 2, 3]));
     }
 }
