--- tests/macros/fn/CurryTest.php
@@ -1,32 +1,0 @@
-<?php
-
-use Baethon\Phln\Phln as P;
-
-class CurryTest extends \PHPUnit\Framework\TestCase
-{
-    public function test_it_curries_function()
-    {
-        $sum = function ($a, $b, $c) {
-            return $a + $b + $c;
-        };
-
-        $expected = $sum(1, 2, 3);
-        $g = P::curry($sum);
-
-        $this->assertEquals($expected, $g(1, 2, 3));
-        $this->assertEquals($expected, $g(1)(2, 3));
-        $this->assertEquals($expected, $g(1, 2)(3));
-        $this->assertEquals($expected, $g(1)(2)(3));
-    }
-
-    function test_it_returns_value_when_arguments_lenght_is_equal_to_parameters()
-    {
-        $sum = function ($a, $b, $c) {
-            return $a + $b + $c;
-        };
-        $expected = $sum(1, 2, 3);
-        $g = P::curry($sum, [1, 2, 3]);
-
-        $this->assertEquals($expected, $g);
-    }
-}
